# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Campus Radio Station Song Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Song {
  id: number;
  title: string;
  artist: string;
  /** @nullable */
  album?: string | null;
  /** @nullable */
  genre?: string | null;
  playedCount: number;
  /** @nullable */
  lastPlayedAt?: string | null;
  /** @nullable */
  lastPlayedBy?: string | null;
  isStudentSubmission: boolean;
  /** @nullable */
  submittedBy?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface SongInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  artist: string;
  album?: string;
  genre?: string;
  isStudentSubmission?: boolean;
  submittedBy?: string;
  notes?: string;
}

export interface SongUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  artist?: string;
  album?: string;
  genre?: string;
  notes?: string;
}

export interface PlayInput {
  djName?: string;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Submission {
  id: number;
  title: string;
  artist: string;
  studentName: string;
  /** @nullable */
  studentEmail?: string | null;
  /** @nullable */
  message?: string | null;
  status: SubmissionStatus;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: string;
}

export interface SubmissionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  artist: string;
  /** @minLength 1 */
  studentName: string;
  studentEmail?: string;
  message?: string;
}

export type SubmissionUpdateStatus = typeof SubmissionUpdateStatus[keyof typeof SubmissionUpdateStatus];


export const SubmissionUpdateStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface SubmissionUpdate {
  status?: SubmissionUpdateStatus;
  reviewNote?: string;
}

export interface StationStats {
  totalSongs: number;
  playedSongs: number;
  unplayedSongs: number;
  totalPlays: number;
  pendingSubmissions: number;
  approvedSubmissions: number;
  rejectedSubmissions: number;
}

export type ListSongsParams = {
status?: ListSongsStatus;
source?: ListSongsSource;
search?: string;
};

export type ListSongsStatus = typeof ListSongsStatus[keyof typeof ListSongsStatus];


export const ListSongsStatus = {
  played: 'played',
  unplayed: 'unplayed',
  all: 'all',
} as const;

export type ListSongsSource = typeof ListSongsSource[keyof typeof ListSongsSource];


export const ListSongsSource = {
  student: 'student',
  staff: 'staff',
  all: 'all',
} as const;

export type ListRecentPlaysParams = {
limit?: number;
};

export type ListSubmissionsParams = {
status?: ListSubmissionsStatus;
};

export type ListSubmissionsStatus = typeof ListSubmissionsStatus[keyof typeof ListSubmissionsStatus];


export const ListSubmissionsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  all: 'all',
} as const;