# lib/api-zod/src/generated/types/songUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Campus Radio Station Song Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SongUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  artist?: string;
  album?: string;
  genre?: string;
  notes?: string;
}