# lib/api-zod/src/generated/types/stationStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Campus Radio Station Song Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface StationStats {
  totalSongs: number;
  playedSongs: number;
  unplayedSongs: number;
  totalPlays: number;
  pendingSubmissions: number;
  approvedSubmissions: number;
  rejectedSubmissions: number;
}