# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Campus Radio Station Song Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionStatus } from './submissionStatus';

export interface Submission {
  id: number;
  title: string;
  artist: string;
  studentName: string;
  /** @nullable */
  studentEmail?: string | null;
  /** @nullable */
  message?: string | null;
  status: SubmissionStatus;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: string;
}