# lib/api-zod/src/generated/types/submissionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Campus Radio Station Song Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmissionInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  artist: string;
  /** @minLength 1 */
  studentName: string;
  studentEmail?: string;
  message?: string;
}