# lib/api-zod/src/generated/types/submissionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Campus Radio Station Song Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmissionUpdateStatus } from './submissionUpdateStatus';

export interface SubmissionUpdate {
  status?: SubmissionUpdateStatus;
  reviewNote?: string;
}